+++ lib/patient_http/client.rb
@@ -12,6 +12,7 @@ module PatientHttp
         protocol: config.protocol
       )
       @response_reader = ResponseReader.new(@processor)
+      @request_preparer = RequestPreparer.new(config)
     end
 
     # Make an asynchronous HTTP request.
@@ -23,8 +24,9 @@ module PatientHttp
       async_response = nil
 
       begin
-        headers = request_headers(request, request_id)
-        url = config.secret_manager.resolve_url(request.url, request.secret_params)
+        outgoing = @request_preparer.prepare(request, request_id)
+        url = outgoing.url
+        headers = outgoing.headers.to_h
         body = Protocol::HTTP::Body::Buffered.wrap([request.body.to_s]) if request.body
         timeout = request.timeout || config.request_timeout
 
@@ -71,13 +73,6 @@ module PatientHttp
       else
         false
       end
-    end
-
-    def request_headers(request, request_id)
-      headers = config.secret_manager.resolve_headers(request.headers.to_h)
-      headers["x-request-id"] = request_id
-      headers["user-agent"] ||= config.user_agent if config.user_agent
-      headers
     end
   end
 end